Organize dashboard file structure
