+++ client/src/components/dashboard/dashboards/AdminDashBoard.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 
 export default class AdminDashBoard extends Component {
   render() {
